+++ srcs/builtins/ft_unset.c
@@ -6,7 +6,7 @@
 /*   By: asanthos <asanthos@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/06/01 06:20:37 by asanthos          #+#    #+#             */
-/*   Updated: 2022/08/24 16:08:27 by asanthos         ###   ########.fr       */
+/*   Updated: 2022/08/24 19:57:26 by asanthos         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 static int	ch_equality(char *str, t_env *lst)
 {
-	t_env *tmp2;
+	t_env	*tmp2;
 
 	if (ft_strcmp(str, lst->name) == 0)
 	{
@@ -31,8 +31,8 @@ static int	ch_equality(char *str, t_env *lst)
 
 void	ft_unset(t_env *lst, t_cmd *cmd_lst)
 {
-	int i;
-	t_env *tmp;
+	size_t	i;
+	t_env	*tmp;
 
 	i = 1;
 	tmp = lst;
@@ -43,7 +43,6 @@ void	ft_unset(t_env *lst, t_cmd *cmd_lst)
 			if (ch_equality(cmd_lst->argument[i], lst) == 1)
 			{
 				lst = tmp;
-				// free_cmd(cmd_lst);
 				return ;
 			}
 			lst = lst->next;
@@ -51,7 +50,6 @@ void	ft_unset(t_env *lst, t_cmd *cmd_lst)
 		if (ch_equality(cmd_lst->argument[i], lst) == 1)
 		{
 			lst = tmp;
-			// free_cmd(cmd_lst);
 			return ;
 		}
 		lst = tmp;
